requirements.txt
alembic==1.10.3
click==8.1.3
Flask==2.2.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
PyJWT==2.6.0
python-dotenv==1.0.0
SQLAlchemy==2.0.9
typing_extensions==4.5.0
Werkzeug==2.2.3
